+ First version of the language file

+++ src/main/java/joserodpt/realpermissions/commands/RealPermissionsCMD.java
@@ -15,6 +15,7 @@ package joserodpt.realpermissions.commands;
 
 import joserodpt.realpermissions.RealPermissions;
 import joserodpt.realpermissions.config.Config;
+import joserodpt.realpermissions.config.Language;
 import joserodpt.realpermissions.config.Players;
 import joserodpt.realpermissions.config.Ranks;
 import joserodpt.realpermissions.gui.RealPermissionsGUI;
@@ -34,6 +35,8 @@ import java.util.Arrays;
 @Command("realpermissions")
 @Alias("rp")
 public class RealPermissionsCMD extends CommandBase {
+
+    private final String noConsole = "[RealPermissions] Only players can run this command.";
 
     RealPermissions rp;
     public RealPermissionsCMD(RealPermissions rp) {
@@ -58,11 +61,12 @@ public class RealPermissionsCMD extends CommandBase {
     @Permission("realpermissions.admin")
     public void reloadcmd(final CommandSender commandSender) {
         Config.reload();
+        Language.reload();
         Ranks.reload();
         rp.getRankManager().loadRanks();
         rp.getRankManager().loadRankups();
         Players.reload();
-        Text.send(commandSender, "&aReloaded!");
+        Text.send(commandSender, Language.file().getString("System.Reloaded"));
     }
 
     @SubCommand("rank")
@@ -75,14 +79,14 @@ public class RealPermissionsCMD extends CommandBase {
 
             Rank r = rp.getRankManager().getRank(rank);
             if (r == null) {
-                Text.send(p, "There is no rank named &c" + rank);
+                Text.send(p, Language.file().getString("Ranks.No-Rank-Found").replace("%name%", rank));
                 return;
             }
 
             RankGUI rg = new RankGUI(p, r, rp);
             rg.openInventory(p);
         } else {
-            commandSender.sendMessage("[RealPermissions] Only players can run this command.");
+            Text.send(commandSender, noConsole);
         }
     }
 
@@ -97,7 +101,7 @@ public class RealPermissionsCMD extends CommandBase {
             PlayersGUI rg = new PlayersGUI(p, rp);
             rg.openInventory(p);
         } else {
-            commandSender.sendMessage("[RealPermissions] Only players can run this command.");
+            Text.send(commandSender, noConsole);
         }
     }
 
@@ -126,12 +130,12 @@ public class RealPermissionsCMD extends CommandBase {
         }
 
         if (p == null) {
-            Text.send(commandSender, "There is no player named like provided.");
+            Text.send(commandSender, Language.file().getString("System.No-Player-Found"));
             return;
         }
 
         rp.getPlayerManager().getPlayer(p).setSuperUser(!rp.getPlayerManager().getPlayer(p).isSuperUser());
-        Text.send(commandSender, p.getName() + "'s &fsuper user: " + (rp.getPlayerManager().getPlayer(p).isSuperUser() ? "&aON" : "&cOFF"));
+        Text.send(commandSender, Language.file().getString("System.Super-User-State").replace("%player%", p.getName()).replace("%state%", (rp.getPlayerManager().getPlayer(p).isSuperUser() ? "&aON" : "&cOFF")));
     }
 
     @SubCommand("set")
@@ -142,24 +146,23 @@ public class RealPermissionsCMD extends CommandBase {
     public void setrankcmd(final CommandSender commandSender, final Player p, final String rank) {
         if (commandSender instanceof Player) {
             if (rp.getPlayerManager().isNotSuperUser((Player) commandSender)) {
-                Text.send(commandSender, " &cYou don't have permission to execute this command!");
+                Text.send(commandSender, Language.file().getString("System.No-Permission-Command"));
                 return;
             }
         }
 
         if (p == null) {
-            Text.send(commandSender, "There is no player named like provided.");
+            Text.send(commandSender, Language.file().getString("System.No-Player-Found"));
             return;
         }
 
         Rank r = rp.getRankManager().getRank(rank);
         if (r == null) {
-            Text.send(commandSender, "There is no rank named &c" + rank);
-            return;
+            Text.send(p, Language.file().getString("Ranks.No-Rank-Found").replace("%name%", rank));            return;
         }
 
         rp.getPlayerManager().getPlayer(p).setRank(r);
-        Text.send(commandSender, p.getName() + "'s &frank is now: " + r.getPrefix());
+        Text.send(commandSender, Language.file().getString("Ranks.Rank-Set").replace("%player%", p.getName()).replace("%rank%", r.getPrefix()));
     }
 
     @SubCommand("settimedrank")
@@ -170,29 +173,28 @@ public class RealPermissionsCMD extends CommandBase {
     public void settimedrankcmd(final CommandSender commandSender, final Player p, final String rank, final Integer seconds) {
         if (commandSender instanceof Player) {
             if (rp.getPlayerManager().isNotSuperUser((Player) commandSender)) {
-                Text.send(commandSender, " &cYou don't have permission to execute this command!");
+                Text.send(commandSender, Language.file().getString("System.No-Permission-Command"));
                 return;
             }
         }
 
         if (p == null) {
-            Text.send(commandSender, "There is no player named like provided.");
+            Text.send(commandSender, Language.file().getString("System.No-Player-Found"));
             return;
         }
 
         Rank r = rp.getRankManager().getRank(rank);
         if (r == null) {
-            Text.send(commandSender, "There is no rank named &c" + rank);
-            return;
+            Text.send(p, Language.file().getString("Ranks.No-Rank-Found").replace("%name%", rank));            return;
         }
 
         if (seconds == null || seconds <= 0) {
-            Text.send(commandSender, "Timed ranks must be set above 0 seconds.");
+            Text.send(commandSender, Language.file().getString("Ranks.Timed-Rank-Above-Zero"));
             return;
         }
 
         rp.getPlayerManager().getPlayer(p).setTimedRank(r, seconds);
-        Text.send(commandSender, p.getName() + "'s &ftimed rank is now: " + r.getPrefix());
+        Text.send(commandSender, Language.file().getString("Ranks.Timed-Rank-Set").replace("%player%", p.getName()).replace("%rank%", r.getPrefix()));
     }
 
     @SubCommand("cleartimedrank")
@@ -203,21 +205,21 @@ public class RealPermissionsCMD extends CommandBase {
     public void cleartimedcmd(final CommandSender commandSender, final Player p) {
         if (commandSender instanceof Player) {
             if (rp.getPlayerManager().isNotSuperUser((Player) commandSender)) {
-                Text.send(commandSender, " &cYou don't have permission to execute this command!");
+                Text.send(commandSender, Language.file().getString("System.No-Permission-Command"));
                 return;
             }
         }
 
         if (p == null) {
-            Text.send(commandSender, "There is no player named like provided.");
+            Text.send(commandSender, Language.file().getString("System.No-Player-Found"));
             return;
         }
 
         if (rp.getPlayerManager().getPlayer(p).hasTimedRank()) {
             rp.getPlayerManager().getPlayer(p).removeTimedRank();
-            Text.send(commandSender, p.getName() + "'s &ftimed rank has been removed.");
+            Text.send(commandSender, Language.file().getString("Ranks.Player-Remove-Timed-Rank").replace("%player%", p.getName()));
         } else {
-            Text.send(commandSender, p.getName() + " doesn't have a timed rank.");
+            Text.send(commandSender, Language.file().getString("Ranks.Player-No-Timed-Rank").replace("%player%", p.getName()));
         }
     }
 
@@ -229,17 +231,16 @@ public class RealPermissionsCMD extends CommandBase {
     public void renamecmd(final CommandSender commandSender, final String rank, final String name) {
         Rank r = rp.getRankManager().getRank(rank);
         if (r == null) {
-            Text.send(commandSender, "There is no rank named &c" + rank);
-            return;
+            Text.send(commandSender, Language.file().getString("Ranks.No-Rank-Found").replace("%name%", rank));            return;
         }
 
         if (name.isEmpty()) {
-            Text.send(commandSender, "New rank name is empty.");
+            Text.send(commandSender, Language.file().getString("Ranks.Name-Empty"));
             return;
         }
 
         rp.getRankManager().renameRank(r, name);
-        Text.send(commandSender, "The rank's name is now " + name);
+        Text.send(commandSender, Language.file().getString("Ranks.New-Name").replace("%name%", name));
     }
 
     @SubCommand("delete")
@@ -250,19 +251,22 @@ public class RealPermissionsCMD extends CommandBase {
     public void delrankcmd(final CommandSender commandSender, final String rank) {
         if (commandSender instanceof Player) {
             if (rp.getPlayerManager().isNotSuperUser((Player) commandSender)) {
-                Text.send(commandSender, " &cYou don't have permission to execute this command!");
+                Text.send(commandSender, Language.file().getString("System.No-Permission-Command"));
                 return;
             }
         }
 
         Rank r = rp.getRankManager().getRank(rank);
         if (r == null) {
-            Text.send(commandSender, "There is no rank named &c" + rank);
-            return;
+            Text.send(commandSender, Language.file().getString("Ranks.No-Rank-Found").replace("%name%", rank));            return;
         }
 
-        rp.getRankManager().deleteRank(r);
-        Text.send(commandSender, r.getPrefix() + " &frank &cdeleted.");
+        if (rp.getRankManager().getDefaultRank() == r) {
+            Text.send(commandSender, Language.file().getString("Ranks.Cant-Delete-Default-Rank"));
+        } else {
+            rp.getRankManager().deleteRank(r);
+            Text.send(commandSender, Language.file().getString("Ranks.Deleted").replace("%rank%", r.getPrefix()));
+        }
     }
 
     @SubCommand("permission")
@@ -273,7 +277,7 @@ public class RealPermissionsCMD extends CommandBase {
     public void permcmd(final CommandSender commandSender, final String operation, final String rank, final String perm) {
         if (commandSender instanceof Player) {
             if (rp.getPlayerManager().isNotSuperUser((Player) commandSender)) {
-                Text.send(commandSender, " &cYou don't have permission to execute this command!");
+                Text.send(commandSender, Language.file().getString("System.No-Permission-Command"));
                 return;
             }
         }
@@ -286,35 +290,34 @@ public class RealPermissionsCMD extends CommandBase {
                 add = false;
                 break;
             default:
-                Text.send(commandSender, "There is no operation like that. Use add/remove.");
+                Text.send(commandSender, "&cInvalid Operation. &fValid Operations: add/remove.");
                 return;
         }
 
         Rank r = rp.getRankManager().getRank(rank);
         if (r == null) {
-            Text.send(commandSender, "There is no rank named &c" + rank);
-            return;
+            Text.send(commandSender, Language.file().getString("Ranks.No-Rank-Found").replace("%name%", rank));            return;
         }
 
         if (add) {
             if (r.hasPermission(perm)) {
-                Text.send(commandSender, "The rank already has the " + perm + " permission.");
+                Text.send(commandSender, Language.file().getString("Permissions.Rank-Already-Has-Permission").replace("%perm%", perm));
             } else {
                 r.addPermission(perm);
                 rp.getRankManager().refreshPermsAndPlayers();
-                Text.send(commandSender, "&fPermission " + perm + " &aadded &fto " + r.getPrefix());
+                Text.send(commandSender, Language.file().getString("Permissions.Rank-Perm-Add").replace("%perm%", perm).replace("%rank%", r.getPrefix()));
             }
         } else {
             if (!r.hasPermission(perm)) {
-                Text.send(commandSender, "The rank doesn't have the " + perm + " permission.");
+                Text.send(commandSender, Language.file().getString("Permissions.Rank-Doesnt-Have-Permission").replace("%perm%", perm));
             } else {
                 joserodpt.realpermissions.permission.Permission p = r.getPermission(perm);
-                if (!p.getAssociatedRank().equalsIgnoreCase(r.getName())) {
-                    Text.send(commandSender, "&fThis permission is associated with the rank: " + p.getAssociatedRank() + ". Remove it in the corresponding rank.");
+                if (!p.getAssociatedRankName().equalsIgnoreCase(r.getName())) {
+                    Text.send(commandSender, Language.file().getString("Permissions.Permission-Associated-With-Other-Rank").replace("%rank%", p.getAssociatedRankName()));
                 } else {
                     r.removePermission(perm);
                     rp.getRankManager().refreshPermsAndPlayers();
-                    Text.send(commandSender, "&fPermission " + perm + " &cremoved &ffrom " + r.getPrefix());
+                    Text.send(commandSender, Language.file().getString("Permissions.Rank-Perm-Remove").replace("%perm%", perm).replace("%rank%", r.getPrefix()));
                 }
             }
         }
@@ -328,7 +331,7 @@ public class RealPermissionsCMD extends CommandBase {
     public void permcmd(final CommandSender commandSender, final String operation, final Player p, final String perm) {
         if (commandSender instanceof Player) {
             if (rp.getPlayerManager().isNotSuperUser((Player) commandSender)) {
-                Text.send(commandSender, " &cYou don't have permission to execute this command!");
+                Text.send(commandSender, Language.file().getString("System.No-Permission-Command"));
                 return;
             }
         }
@@ -341,12 +344,12 @@ public class RealPermissionsCMD extends CommandBase {
                 add = false;
                 break;
             default:
-                Text.send(commandSender, "There is no operation like that. Use add/remove.");
+                Text.send(commandSender, "&cInvalid Operation. &fValid Operations: add/remove.");
                 return;
         }
 
         if (p == null) {
-            Text.send(commandSender, "There is no player named like provided.");
+            Text.send(commandSender, Language.file().getString("System.No-Player-Found"));
             return;
         }
 
@@ -354,19 +357,17 @@ public class RealPermissionsCMD extends CommandBase {
 
         if (add) {
             if (pa.hasPermission(perm)) {
-                Text.send(commandSender, "The player already has the " + perm + " permission.");
+                Text.send(commandSender, Language.file().getString("Permissions.Player.Already-Has-Permission"));
             } else {
                 pa.addPermission(perm);
-                //                    rp.getRankManager().refreshPermsAndPlayers();
-                Text.send(commandSender, "&fPermission " + perm + " &aadded &fto " + p.getName());
+                Text.send(commandSender, Language.file().getString("Permissions.Player.Add").replace("%perm%", perm).replace("%player%", p.getName()));
             }
         } else {
             if (!pa.hasPermission(perm)) {
-                Text.send(commandSender, "The player doesn't have the " + perm + " permission.");
+                Text.send(commandSender, Language.file().getString("Permissions.Player.Doesnt-Have-Permission"));
             } else {
                 pa.removePermission(perm);
-                //rp.getRankManager().refreshPermsAndPlayers();
-                Text.send(commandSender, "&fPermission " + perm + " &cremoved &ffrom " + p.getName());
+                Text.send(commandSender, Language.file().getString("Permissions.Player.Remove").replace("%perm%", perm).replace("%player%", p.getName()));
             }
         }
     }
